apps/sta/src/app/feature/auto-tagging/text-input/text-input.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { ReactiveFormsModule } from '@angular/forms';
import { MatButtonModule } from '@angular/material/button';
import { MatFormFieldModule } from '@angular/material/form-field';
import { MatInputModule } from '@angular/material/input';
import { NoopAnimationsModule } from '@angular/platform-browser/animations';

import { configureTestSuite } from 'ng-bullet';

import { DataService } from '../../../shared/result/data.service';
import { TextInputComponent } from './text-input.component';

describe('TextInputComponent', () => {
  let component: TextInputComponent;
  let fixture: ComponentFixture<TextInputComponent>;
  let dataService: DataService;

  configureTestSuite(() => {
    TestBed.configureTestingModule({
      imports: [
        NoopAnimationsModule,
        MatFormFieldModule,
        MatInputModule,
        ReactiveFormsModule,
        MatButtonModule
      ],
      declarations: [TextInputComponent],
      providers: [{ provide: DataService, useValue: {} }]
    });
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(TextInputComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();

    dataService = TestBed.get(DataService);
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  describe('ngOnInit', () => {
    test('it should init form control', () => {
      // tslint:disable-next-line: no-lifecycle-call
      component.ngOnInit();

      expect(component.textFormControl).toBeDefined();
    });
  });

  describe('getTags()', () => {
    test('should call postTaggingText from Service', () => {
      dataService.postTaggingText = jest.fn();
      component.textFormControl.setValue('test');

      component.getTags();

      expect(dataService.postTaggingText).toHaveBeenCalledWith('test');
    });
  });
});
